[可能是优化]爬榜

+++ esmwiki-common/src/main/java/moe/zr/service/EventService.java
@@ -7,12 +7,13 @@ import javax.crypto.IllegalBlockSizeException;
 import java.io.IOException;
 import java.text.ParseException;
 import java.util.concurrent.ExecutionException;
+import java.util.concurrent.Future;
 import java.util.concurrent.TimeoutException;
 
 public interface EventService {
 
-    ListenableFuture<Integer> saveAllPointRanking() throws BadPaddingException, InterruptedException, ParseException, IOException, ExecutionException, IllegalBlockSizeException, TimeoutException;
+    ListenableFuture<Long> saveAllPointRanking() throws BadPaddingException, InterruptedException, ParseException, IOException, ExecutionException, IllegalBlockSizeException, TimeoutException;
 
-    ListenableFuture<Integer> saveAllScoreRanking() throws BadPaddingException, InterruptedException, ParseException, IOException, ExecutionException, IllegalBlockSizeException, TimeoutException;
+    ListenableFuture<Long> saveAllScoreRanking() throws BadPaddingException, InterruptedException, ParseException, IOException, ExecutionException, IllegalBlockSizeException, TimeoutException;
 
 }
